refactor(server): route 走行を別 TU (mapoi_nav2_bridge_route.cpp) へ分離
#345 第1弾。FollowWaypoints/mapoi 主導 waypoint 到達のメソッド定義を
別ファイルへ移すのみで、クラス構造・挙動は変更しない。
