+++ mapoi_server/CMakeLists.txt
@@ -22,6 +22,7 @@ target_link_libraries(mapoi_server ${_mapoi_yaml_cpp_target})
 
 ament_auto_add_executable(mapoi_nav2_bridge
   src/mapoi_nav2_bridge.cpp
+  src/mapoi_nav2_bridge_route.cpp
 )
 
 ament_auto_add_executable(mapoi_amcl_localization_bridge
@@ -83,6 +84,7 @@ if(BUILD_TESTING)
   ament_auto_add_gtest(test_nav2_bridge_unit
     test/test_nav2_bridge_unit.cpp
     src/mapoi_nav2_bridge.cpp
+    src/mapoi_nav2_bridge_route.cpp
   )
   target_compile_definitions(test_nav2_bridge_unit PRIVATE UNIT_TEST)
